+++ features/employee/components/employee-card.tsx
@@ -1,8 +1,8 @@
 'use client';
 
 import { Employee, getDepartmentLabel } from '@/types/employee';
-import { Card, CardContent } from '@/components/ui/card';
-import { Badge } from '@/components/ui/badge';
+import { Card, CardContent } from '@/components/shadcn/card';
+import { Badge } from '@/components/shadcn/badge';
 import { Mail, Phone, Building, IdCard } from 'lucide-react';
 import Link from 'next/link';
 import { format } from 'date-fns';
